buffer input and output separately
Though having the single buffer looked like it would be more efficient,
in practice, just using bufio's buffers on the input and output proved
to be even more performant - since more bytes are written per Write() -
and greatly simplified the code.

The downside is it uses another 4kb of memory which is probably fine.

Since we are passing data between the buffers (mostly) byte-by-byte, one
improvement would be to batch process this transfer. However, profiling
shows that it is not really a bottleneck.