avoid using temporary files
Mainly since when I used temporary files, every time I tried to execute
them with `sh -c` they would fail with code 126. But also it is a
security boon as well.

Now I simply pass the entire contents of the script as the first
argument of `sh -c`.
Though now the script will potentially show up in the process list for
hackers to read.