Write script to file to execute
This is primarily so that we can use the shebang to customise the
program used to run the application later on.

There is potentially security issues with this where an attacker could
inject things into the file before it is run. However we set the file
permissions to 0700 so anyone that could execute the file already has
user permissions. There could still ways around this though.