Pass positional arguments when reading stdin
We want to be able to read the template from stdin and pass positional
arguments.
This was not possible before because it would treat the first argument
provided to it as an input file. Here we have used the special character
"-" to express the intent that the the template should be read from
stdin, and so we are free to pass positional arguments after it.