Execute the code block as a positional arg
Pass the code block as a positional argument to the shell program rather
than in a temporary file.

This has the advantage of not having to clean up temporary files in the
case of panics or the process being killed.

Disadvantages
* Linux normally allows 32 pages for positional arguments in execv(1).
  Code blocks shouldn't really be this large and if the user intends on
  doing so, they can always put the code in a separate file to be called
  from the code block.
* The contents of the codeblock are visible in the process list
* Commands used to execute the codeblock must be able to take the script
  in the `-c` flag e.g. sh -c "echo Hello". This covers most shells and
  python which is the vast majority of usecases. Otherwise the user
  could write a script that wraps around their particular program to
  suit this format.

Overall, the program becomes slightly less flexible (though with
workarounds) in order to simplify the code whilst satisfying 99% of the
usecases.